BigDecimal

+++ src/main/java/com/itwillbs/domain/MaterialReservationVO.java
@@ -1,5 +1,6 @@
 package com.itwillbs.domain;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import lombok.Data;
@@ -14,7 +15,7 @@ public class MaterialReservationVO {
 	
 	private String workOrderId;			// 작업지시서 번호
 	private String materialId;			// 자재 ID
-	private int reservedQty;			// 예약 수량 (해당 작업지시서가 확보해둔 수량)
+	private BigDecimal reservedQty;		// 예약 수량 (해당 작업지시서가 확보해둔 수량)
 	private Date updatedAt;				// 수정 시간
 
 }
